backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# Auth
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
python-multipart==0.0.20

# Validation & Settings
pydantic[email]==2.10.4
pydantic-settings==2.7.1

# Date/time utilities
python-dateutil==2.9.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1

# Dev
python-dotenv==1.0.1

# AI (multi-provider: Claude, OpenAI, xAI/Grok)
anthropic>=0.40.0
openai>=1.0.0

# Email
aiosmtplib>=3.0.0

# Excel export
openpyxl>=3.1.0
